requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=2.0 # Apache-2.0
openstacksdk>=0.12.0 # Apache-2.0
oslo.config>=5.2.0 # Apache-2.0
ply>=3.10 # BSD 3 Clauses
z3-solver>=4.8.0 # MIT
prettytable>=0.7.2 #BSD 3 Clauses
python-neutronclient>=6.8.0  # Apache-2.0
skydive-client>=0.5 # Apache-2.0

